requirements.txt
diango~=1.11
gunicorn==19.9.1
supervisor==4.0.0dev0
xadmin==0.6.1
mysqlclient==1.3.12
django-ckeditor==5.4.0
django-rest-framework==0.1.0
django-redis==4.8.0
django-autocomplete=light==3.2.10
mistune==0.8.3
django-debug-toolbar==1.9.1